Implement dropout
Fix some reshape handling to avoid changing the behaviour of built-in
ops (where the result depends on the shape).

Make use of push constants for now to change the random seed for each
run of the graph (might replace with proper uniform variables in
future).  Dropout behaviour is intended to be deterministic!

Change per-element clusters so that only per-element ops are members
of the cluster (literals and built-ins behave like loads).