Refactor layers API to simpler building blocks
Tag variables with a boolean to specify if they are trainable or not.

Add a method to the graph API to gather all trainable parameters that
are referenced by the graph, instead of having to keep track of them
separately.

Simplify layers into an API where the caller stores an instance of some
building block (which owns the parameters for that block), and manually
calls it to emit ops into a graph.