Allow gather loads to fuse with other ops
Allow the load index and result to be fused with other per-element ops.
The graph edge into the values is considered as a "gather" edge and
is not allowed to be fused (we must force values to be written to
memory so we can index into them).

Removed the special kernel for gather loads, not needed any more.